--- news_website/templates/user_profile.html
@@ -1,60 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>CNN News</title>
-    <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
-</head>
-<body>
-    <div class="background"></div>
-    <div class="container1">
-        <div class="clicked_news">
-            <div>
-                <a href="/" type="button" class="btn btn-light return">返回</a>
-            </div>
-            <table class="table caption-top table-bordered border-dark">
-                <caption><h4>{{ user }}</h4></caption>
-                <tr>
-                    <th colspan="4" class="clicked_title">
-                        <strong>Clicked News</strong>
-                    </th>
-                </tr>
-                <tr>
-                    <th class="table_text">#</th>
-                    <th class="table_text">ID</th>
-                    <th class="table_text">Category</th>
-                    <th class="table_text">Title</th>
-                </tr>
-                {% for index, news in clicked_articles.iterrows() %}
-                    <tr class="table_text">
-                        <th class="table_text">{{ news['id'] }}</th>
-                        <th class="table_text">{{ news.name }}</th>
-                        <th class="table_text">{{ news['category'] }}</th>
-                        <th class="table_text">{{ news['title'] }}</th>
-                    </tr>
-                {% endfor %}
-            </table>
-        </div>
-        <div class="candidate_news">
-            <img src="{{ url_for('wordcloud_image') }}" alt="Word Cloud">
-            <table class="table caption-top table-bordered border-dark">
-                <tr>
-                    <th colspan="4" class="candidate_title">
-                        <strong>Top5 News Recommended by {{ user }}</strong>
-                    </th>
-                </tr>
-                {% for index, news in candidate_articles.iterrows() %}
-                    <tr>
-                        <th class="table_text {{ news['color'] }}">{{ news.name }}</th>
-                        <th class="table_text {{ news['color'] }}">{{ news['category'] }}</th>
-                        <th class="table_text {{ news['color'] }}">{{ news['title'] }}</th>
-                    </tr>
-                {% endfor %}
-            </table>
-        </div>
-    </div>
-</body>
-</html>
